requirements.txt
Pillow
requests
google-generativeai
langchain
langchain_groq 
python-dotenv
fastapi
uvicorn
python-multipart